requirements.txt
discord.py>=2.3.0
google-genai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
google-api-python-client>=2.110.0